+++ examples/counter/src/app/app.component.ts
@@ -2,20 +2,12 @@ import { Component } from '@angular/core';
 import { NgRedux, DevToolsExtension } from 'ng2-redux';
 import { IAppState, rootReducer, enhancers } from '../store/index';
 const createLogger = require('redux-logger');
-import thunk from 'redux-thunk';
 
 @Component({
-  selector: 'root',
-  template: `
-  <h3>Counter example</h3>
-  <counter></counter>
-  <counter-info></counter-info>
-
-  <h3>String length example</h3>
-  <search></search>
-  `
+  selector: 'app-root',
+  templateUrl: './app.component.html',
 })
-export class App {
+export class AppComponent {
   constructor(
     private ngRedux: NgRedux<IAppState>,
     private devTool: DevToolsExtension) {
@@ -23,7 +15,7 @@ export class App {
     this.ngRedux.configureStore(
       rootReducer,
       {},
-      [ createLogger(), thunk ],
+      [ createLogger() ],
       [ ...enhancers, devTool.isEnabled() ? devTool.enhancer() : f => f]);
   }
 }