+++ CHANGELOG.md
@@ -1,3 +1,63 @@
+# 5.0.0
+
+* Fix for the `ERROR in NgReduxModule is not an NgModule` error thrown by Angular CLI.
+* Remove deprecations.
+* Breaking changes associated with Angular 2.4+.
+
+### Breaking Changes
+
+* Minimum Angular peer dependency is now 2.4.0
+* Removed support for the `connect` pattern: it's simply not a good fit for Angular.
+You should be using the `select` pattern now.
+* Remove deprecated constructor arg for `NgRedux`.
+* `NgReduxModule.forRoot` is no more, due to Angular API changes. Now just import `NgReduxModule` directly.
+
+#### Old Way:
+
+```typescript
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { NgReduxModule } from 'ng2-redux';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    NgReduxModule.forRoot(),
+    BrowserModule,
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+class AppModule {
+  // etc.
+}
+```
+
+#### New Way:
+
+```typescript
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { NgReduxModule } from 'ng2-redux';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    NgReduxModule,
+    BrowserModule,
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+class AppModule {
+  // etc.
+}
+```
+
 # 4.2.4
 
 Recovery release that restores the functionality of 4.2.2. Use this release
@@ -57,7 +117,7 @@ import { rootReducer } from './store';
     AppComponent
   ],
   imports: [
-    NgReduxModule.forRoot(),
+    NgReduxModule,
     BrowserModule,
   ],
   providers: [],